Check cluster health before scheduling queries (#14)
This PR adds functionality to check that a Trino cluster is healthy
before scheduling queries to it. Scheduling a query when the server is
initializing will result in a failed query. The Trino docs say that the
server should return a 500/503 when this is happening, but I couldn't
replicate that functionality. It appears that the http server starts
accepting queries (and returning 200s) before they're able to be queued.

In the case that there are no enabled or healthy clusters, a DD metric
with the query backlog is incremented.

In addition, the level of some logs was reduced to prevent noisy
messages while developing and in QA environments.