Update cluster patch route to not require all keys (#29)
This change modifies the patch route to not require all keys when
updating a cluster. Not a large function change, mostly a quality of
life improvement when updating information about a cluster.

Requests to disable a cluster now only need to include the updated
status:

![image](https://github.com/rockettools/trino-proxy/assets/26007775/1ffb6ea1-7e49-4416-9d53-b8963a92b838)


Extra keys included in the request body are ignored.