Adding Trino Config Files for Local Testing (#37)
This PR adds in configuration files for trino to allow for custom
settings. This is necessary to set the web-ui authentication type to
"fixed", which is currently necessary to test LOAD based routing
locally. These files can also be adjusted as needed if you want to test
additional Trino configurations.

With the exception of the web-ui authentication setting and adding an
additional config to connect to the proxy DB, these files were copied
from the current latest trino docker image with not other modifications.

PR should be low risk, as there is no actual functionality change to the
proxy itself.