chooseCluster(): check tags only if present (#41)
Prior to #39, when choosing onto which cluster to schedule a new query,
`queryClusterTags` was checked against the Cluster's tags only if the
`clusterHeaderRegex` pattern (`/-- Cluster: *(.*)/`) matched the
`query.body`.

This doesn't really work if most `query.body` payloads don't contain a
tag but all `availableClusters` do: change this behavior to only enforce
tags if they are present. In contrast to #40, this makes it so the
`userClusterTags` are still checked and enforced, such that a user
cannot gain access to a cluster they would not normally have access to
with a leading `-- Cluster: XXX` comment.